+++ json_annotation/lib/src/json_serializable.g.dart
@@ -21,6 +21,7 @@ JsonSerializable _$JsonSerializableFromJson(
         'create_factory',
         'create_field_map',
         'create_json_keys',
+        'create_json_schema',
         'create_per_field_to_json',
         'create_to_json',
         'disallow_unrecognized_keys',
@@ -37,6 +38,14 @@ JsonSerializable _$JsonSerializableFromJson(
       constructor: $checkedConvert('constructor', (v) => v as String?),
       createFieldMap: $checkedConvert('create_field_map', (v) => v as bool?),
       createJsonKeys: $checkedConvert('create_json_keys', (v) => v as bool?),
+      createJsonSchema: $checkedConvert(
+        'create_json_schema',
+        (v) => v as bool?,
+      ),
+      createPerFieldToJson: $checkedConvert(
+        'create_per_field_to_json',
+        (v) => v as bool?,
+      ),
       createFactory: $checkedConvert('create_factory', (v) => v as bool?),
       createToJson: $checkedConvert('create_to_json', (v) => v as bool?),
       disallowUnrecognizedKeys: $checkedConvert(
@@ -57,10 +66,6 @@ JsonSerializable _$JsonSerializableFromJson(
         'generic_argument_factories',
         (v) => v as bool?,
       ),
-      createPerFieldToJson: $checkedConvert(
-        'create_per_field_to_json',
-        (v) => v as bool?,
-      ),
     );
     return val;
   },
@@ -68,6 +73,8 @@ JsonSerializable _$JsonSerializableFromJson(
     'anyMap': 'any_map',
     'createFieldMap': 'create_field_map',
     'createJsonKeys': 'create_json_keys',
+    'createJsonSchema': 'create_json_schema',
+    'createPerFieldToJson': 'create_per_field_to_json',
     'createFactory': 'create_factory',
     'createToJson': 'create_to_json',
     'disallowUnrecognizedKeys': 'disallow_unrecognized_keys',
@@ -76,7 +83,6 @@ JsonSerializable _$JsonSerializableFromJson(
     'ignoreUnannotated': 'ignore_unannotated',
     'includeIfNull': 'include_if_null',
     'genericArgumentFactories': 'generic_argument_factories',
-    'createPerFieldToJson': 'create_per_field_to_json',
   },
 );
 
@@ -88,6 +94,7 @@ Map<String, dynamic> _$JsonSerializableToJson(JsonSerializable instance) =>
       'create_factory': instance.createFactory,
       'create_field_map': instance.createFieldMap,
       'create_json_keys': instance.createJsonKeys,
+      'create_json_schema': instance.createJsonSchema,
       'create_per_field_to_json': instance.createPerFieldToJson,
       'create_to_json': instance.createToJson,
       'disallow_unrecognized_keys': instance.disallowUnrecognizedKeys,