+++ json_serializable/example/example.dart
@@ -6,7 +6,7 @@ import 'package:json_annotation/json_annotation.dart';
 
 part 'example.g.dart';
 
-@JsonSerializable()
+@JsonSerializable(createJsonSchema: true)
 class Person {
   /// The generated code assumes these values exist in JSON.
   final String firstName, lastName;
@@ -23,4 +23,7 @@ class Person {
 
   /// Connect the generated [_$PersonToJson] function to the `toJson` method.
   Map<String, dynamic> toJson() => _$PersonToJson(this);
+
+  /// The JSON Schema for this class.
+  static const jsonSchema = _$PersonJsonSchema;
 }