+++ json_serializable/example/example.g.dart
@@ -21,3 +21,25 @@ Map<String, dynamic> _$PersonToJson(Person instance) => <String, dynamic>{
   'lastName': instance.lastName,
   'dateOfBirth': instance.dateOfBirth?.toIso8601String(),
 };
+
+const _$PersonJsonSchema = {
+  r'$schema': 'https://json-schema.org/draft/2020-12/schema',
+  'type': 'object',
+  'properties': {
+    'firstName': {
+      'type': 'string',
+      'description': 'The generated code assumes these values exist in JSON.',
+    },
+    'lastName': {
+      'type': 'string',
+      'description': 'The generated code assumes these values exist in JSON.',
+    },
+    'dateOfBirth': {
+      'type': 'string',
+      'format': 'date-time',
+      'description':
+          "The generated code below handles if the corresponding JSON value doesn't\nexist or is empty.",
+    },
+  },
+  'required': ['firstName', 'lastName'],
+};