+++ json_annotation/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.9.1-wip
 
 - Support `JsonKey` annotation on constructor parameters.
+- Add `JsonSerializable(createJsonSchema: true)`.
 - Require `meta: ^1.16.0`
 - Require `sdk: ^3.9.0`